+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # LuauPolyfills Changelog
 
-# Unreleased
+# 0.2.1
 
 ### Added Polyfills
 * `String.split`